REST APIs- The Practical Application complete

+++ controllers/feed.js
@@ -19,9 +19,7 @@ exports.getPosts = (req, res, next) => {
         .limit(perPage);
     })
     .then(posts => {
-      res
-        .status(200)
-        .json({ 
+      res.status(200).json({ 
           message: 'Fetched posts successfully.', 
           posts: posts,
           totalItems: totalItems
@@ -52,11 +50,12 @@ exports.createPost = (req, res, next) => {
   const content = req.body.content;
   let creator;
   const post = new Post({
-    title: title, 
-    content: content, 
+    title: title,
+    content: content,
     imageUrl: imageUrl,
     creator: req.userId
   });
+  console.log(post);
   post
     .save()
     .then(result => {
@@ -127,6 +126,11 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized!');
+        error.statusCode = 403;
+        throw error;
+      }
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -155,12 +159,23 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized!');
+        error.statusCode = 403;
+        throw error;
+      }
       // Check logged in user
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId);
     })
     .then(result => {
-      console.log(result);
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      user.posts.pull(postId);
+      return user.save();
+    })
+    .then(result => {
       res.status(200).json({ message: 'Deleted post.' });
     })
     .catch(err => {
